requirements.txt
# Core data science stack
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3

# Visualization and notebooks
matplotlib>=3.7
seaborn>=0.12
jupyterlab>=4.0

# Utilities
PyYAML>=6.0
joblib>=1.2

# Modeling extras
xgboost>=1.7

# Data sourcing
kaggle
python-dotenv

# GUI
streamlit>=1.33
